requirements.txt
dash
plotly
dash-bootstrap-components
dash-bootstrap-templates
dash-mantine-components
Pillow
pandas
black
pylint
pytest
dash-auth
s3fs
scipy